cast index to int to avoid formatting as char on some platforms on "[PyTorch] Narrow Device to 2 bytes by narrowing DeviceType and DeviceIndex"
DeviceType pretty clearly only needs 1 byte. DeviceIndex only needs 1 byte given that machines don't have anywhere near 255 GPUs in them as far as I know.

Differential Revision: [D24605460](https://our.internmc.facebook.com/intern/diff/D24605460/)

[ghstack-poisoned]

+++ c10/core/Device.h
@@ -94,9 +94,9 @@ struct C10_API Device final {
   DeviceIndex index_ = -1;
   void validate() {
     TORCH_CHECK(index_ == -1 || index_ >= 0,
-        "Device index must be -1 or non-negative, got ", index_);
+        "Device index must be -1 or non-negative, got ", (int)index_);
     TORCH_CHECK(!is_cpu() || index_ <= 0,
-        "CPU device index must be -1 or zero, got ", index_);
+        "CPU device index must be -1 or zero, got ", (int)index_);
   }
 };
